Checked in By: Bhavesh Nechlani Reason: Code changes for update pending order and cancel pending order.

+++ src/main/java/com/traders/portfolio/domain/Stock.java
@@ -54,7 +54,7 @@ public class Stock extends AbstractAuditingEntity<Long> implements Serializable 
     private String strike;
 
     @Column(name = "lot_size")
-    private float lotSize;
+    private Double lotSize;
 
     @Column(name = "expiry", nullable = false)
     @Temporal(TemporalType.DATE)
